LayoutTests/inspector/canvas/console-record-offscreen-canvas-2d.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="resources/recording-utilities.js"></script>
<script src="resources/recording-2d.js"></script>
<script>
function test() {
    let suite = InspectorTest.createAsyncSuite("Canvas.recordingOffscreen2D");

    suite.addTestCase({
        name: "Canvas.recordingOffscreen2D.Console",
        description: "Check that a recording can be triggered by console.record().",
        test(resolve, reject) {
            consoleRecord(WI.Canvas.ContextType.OffscreenCanvas2D, resolve, reject);
        },
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="load({offscreen: true})">
    <p>Test that CanvasManager is able to record actions made to offscreen 2D canvas contexts.</p>
</body>
</html>
