LayoutTests/inspector/canvas/console-record-offscreen-canvas-2d-expected.txt

Test that CanvasManager is able to record actions made to offscreen 2D canvas contexts.


== Running test suite: Canvas.recordingOffscreen2D
-- Running test case: Canvas.recordingOffscreen2D.Console
PASS: The recording should have the name "TEST".
PASS: The recording should have one frame.
PASS: The first frame should have one action.
